e2e-test/pyproject.toml
[project]
name = "meeplesshelf-e2e"
version = "0.1.0"
description = "End-to-end tests for MeeplesShelf"
requires-python = ">=3.12"
dependencies = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
